backend/requirements.txt
pandas==2.1.3
fastapi==0.104.1
pydantic==2.5.2
optuna == 3.5.0
uvicorn==0.24.0
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.4
PyYAML==6.0.1
lightgbm==3.3.5
scipy==1.11.2
transliterate==1.10.2
geopy==2.4.0
DateTime==5.2
python-multipart==0.0.6